requirements.txt
sqlalchemy
plexapi
pyyaml